requirements.txt
django==2.1.2
psycopg2-binary==2.7.5
djangorestframework==3.8.2